added google auth and made some pages and finalize the whole website

+++ src/app/api/admin/course/[id]/route.ts
@@ -1,39 +1,57 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "../../../../../../prisma/client";
-import { imageKit } from "@/lib/imageKit.js"
+import { imageKit } from "@/lib/imageKit.js";
 import { ObjectId } from "mongodb";
+import slugify from "slugify";
 
 export async function GET(Request: NextRequest, context: { params: Promise<{ id: string }> }) {
   try {
     const { id } = await context.params;
     const course = await prisma.course.findUnique({
-      where: { id: id },
+      where: { id },
       include: { videos: true }
-    })
+    });
     return NextResponse.json(course);
   } catch (error) {
-    console.log(error)
-    return NextResponse.json({ message: "Error fetching course" }, { status: 500 })
+    console.log(error);
+    return NextResponse.json({ message: "Error fetching course" }, { status: 500 });
   }
 }
 
 export async function DELETE(Request: NextRequest, context: { params: Promise<{ id: string }> }) {
   try {
     const { id } = await context.params;
+
+    if (!ObjectId.isValid(id)) {
+      return NextResponse.json({ error: "Invalid course ID" }, { status: 400 });
+    }
+
+    // 1️⃣ Delete all videos first
+    await prisma.video.deleteMany({
+      where: { courseId: id },
+    });
+
+    // 2️⃣ Now delete the course
     const course = await prisma.course.delete({
-      where: { id: id }
-    })
-    return NextResponse.json(course);
+      where: { id },
+    });
+
+    return NextResponse.json({
+      message: "Course deleted successfully",
+      course,
+    });
+
   } catch (error) {
-    console.log(error)
+    console.log("DELETE ERROR:", error);
+    return NextResponse.json(
+      { error: "Failed to delete course" },
+      { status: 500 }
+    );
   }
 }
 
 
-export async function PUT(
-  req: NextRequest,
-  context: { params: Promise<{ id: string }> }
-) {
+export async function PUT(req: NextRequest, context: { params: Promise<{ id: string }> }) {
   try {
     const { id } = await context.params;
 
@@ -42,7 +60,9 @@ export async function PUT(
     }
 
     const body = await req.json();
-    const { imageBase64, id: _, createdAt, updatedAt, ...rest } = body;
+
+  
+    const { imageBase64, id: _, createdAt, updatedAt, videos, slug, ...rest } = body;
 
     let imageUrl = rest.imageUrl;
 
@@ -55,11 +75,17 @@ export async function PUT(
       imageUrl = uploadResponse.url;
     }
 
+    
+    const updatedSlug = rest.title
+      ? slugify(rest.title, { lower: true, strict: true })
+      : undefined;
+
     const course = await prisma.course.update({
       where: { id },
       data: {
         ...rest,
         imageUrl,
+        ...(updatedSlug && { slug: updatedSlug }),
       },
     });
